Improve custom node library loading
Custom node libraries can now be written as ESModules by providing your nodes in
a module that exports a `registerCustomNodes(register: (nodes: CustomNodeInfo[])
=> void): Promise<void> | void;` function. This allows VisualBlocks to know when
all the nodes are done being registered.

For commonjs node libraries, there is a
new function `visualblocks.registerCustomNodes` that registers multiple nodes at
a time. Calling this function will notify VisualBlocks that all your nodes are
loaded.

The original custom node API that uses `visualblocks.registerCustomNode` to
register nodes one at a time remains as it was, but there is now a second
argument that can be passed to it. This argument is called `isLastNode` and it
is `true` by default. Passing `false` will prevent VisualBlocks from assuming
that all the nodes are registered. You must pass `true` on your last call to it.